moved auth and product contrellers in Api folder
